Apparently working now

+++ lib/Interfaiss.ml
@@ -28,7 +28,7 @@ include (
         (* This might happen if the definition of index_t get out of synchro with the C interface *)
         assert false
     type vectors_t = (float, Bigarray.float32_elt, Bigarray.c_layout) Bigarray.Array2.t
-    type offsets_t = (int, Bigarray.int64_elt, Bigarray.c_layout) Bigarray.Array2.t
+    type offsets_t = (int64, Bigarray.int64_elt, Bigarray.c_layout) Bigarray.Array2.t
     external add: t -> vectors_t -> unit = "InterfaissAdd"
     external train: t -> vectors_t -> unit = "InterfaissTrain"
     external query: t -> vectors_t -> int -> vectors_t * offsets_t = "InterfaissQuery"
@@ -41,7 +41,7 @@ include (
     type t
     val create: ?index_type:index_t -> int -> t
     type vectors_t = (float, Bigarray.float32_elt, Bigarray.c_layout) Bigarray.Array2.t
-    type offsets_t = (int, Bigarray.int64_elt, Bigarray.c_layout) Bigarray.Array2.t
+    type offsets_t = (int64, Bigarray.int64_elt, Bigarray.c_layout) Bigarray.Array2.t
     val add: t -> vectors_t -> unit
     val train: t -> vectors_t -> unit
     val query: t -> vectors_t -> int -> vectors_t * offsets_t
